[DebugInfo] Place static variable DIEs under the correct parent
If a static variable (which is also a global variable) is defined in a function that is inlined and removed, then the concrete DIE of the function will not be created and the variable's DIE won't have the correct parent. Instead, we delay assigning the parent DIE until we are sure the function's DIE was created.

Differential Revision: https://reviews.llvm.org/D108492