[LoopRolling] Create pass to roll loops for codesize
This optimization analyzes basic blocks and attempts to roll code into a
loop. You can think of this as the inverse of LoopUnrolling. While this
could degrade runtime performance, it can be useful when users want to
aggresively optimize for codesize.

This was inspired by the paper "Loop Rolling for Code Size Reduction"
(https://doi.org/10.1109/CGO53902.2022.9741256). Thanks to @GiangNPham
who created an initial implementation during an internship.